Moved ObjectTab to commons/ui and fixed Shares tab
