+++ mwdb/web/src/components/ShowObject/index.js
@@ -1,5 +1,4 @@
 export { default as ShowObject } from "./ShowObject";
-export { ObjectTab, ObjectAction, TabContext, useTabContext } from "./ObjectTab";
 
 export { default as RelationsTab } from "./RelationsTab";
 export { default as LatestConfigTab } from "./LatestConfigTab";
@@ -10,3 +9,4 @@ export { default as RemoveAction } from "./RemoveAction";
 export { default as UploadChildAction } from "./UploadChildAction";
 
 export { ObjectContext } from "@mwdb-web/commons/context";
+export { ObjectTab, ObjectAction, TabContext, useTabContext } from "@mwdb-web/commons/ui";
